+++ db4e/Constants/DPane.py
@@ -31,7 +31,7 @@ class DPane(ConstGroup):
     RESULTS: str  = "ResultsPane"
     RUNTIME_LOG: str  = "RuntimeLogPane"
     TUI_LOG: str  = "TuiLogPane"
-    UPTIME: str  = "UptimePane"
+    START_STOP_LOG: str  = "StartStopLogPane"
     WELCOME: str  = "WelcomePane"
     XMRIG: str  = "XMRigPane"
     XMRIG_HASHRATES: str  = "XMRigHashratesPane"