+++ db4e/Panes/P2PoolHashratesPane.py
@@ -29,7 +29,7 @@ class P2PoolHashratesPane(Container):
     instance_label = Label("", id=DForm.INSTANCE_LABEL,classes=DForm.STATIC)
     hashrate_label = Label("", id=DForm.HASHRATE_LABEL, classes=DForm.STATIC)
     hashrate_plot = HashratePlot(DLabel.HASHRATE, id=DField.HASHRATE_PLOT)
-    select_widget = Select(compact=True, id=DForm.TIMES, options=DSelect.SELECT_LIST)
+    select_widget = Select(compact=True, id=DForm.TIMES, options=DSelect.HOURS_SELECT_LIST)
 
 
     def compose(self):